shifted edit user pages to diff components *NOT WORKING*
